pyproject.toml
[tool.poetry]
name = "mscproject"
version = "0.1.0"
description = ""
authors = ["Dominic Thorn <dominic.thorn@gmail.com>"]

[tool.poetry.dependencies]
python = "~=3.9,<3.11"
numpy = "^1.22.4"
pandas = "^1.4.2"
scikit-learn = "^1.1.1"
matplotlib = "^3.5.2"
torch = "^1.11.0"
optuna = "^2.10.0"
mlflow = "^1.26.0"
networkx = "^2.8.2"
igraph = "^0.9.10"
jupyter = "^1.0.0"
pydot = "^1.4.2"
pyspark = "~3.1"
PyYAML = "^6.0"
torch-scatter = { version = "^2.0.9", optional = true }
torch-sparse = { version = "^0.6.13", optional = true }
torch-cluster = { version = "^1.6.0", optional = true }
torch-spline-conv = { version = "^1.2.1", optional = true }
torch-geometric = { version = "^2.1.0", optional = true }
cairocffi = "^1.3.0"
graphframes = "^0.6"
jupytext = "^1.13.8"
graphgym = "^0.4.0"
tensorflow = "^2.9.1"
pandas-downcast = "^1.2.1"
scipy = "^1.8.1"
catboost = "^1.0.6"
shap = "^0.41.0"
tensorflow-gnn = "^0.2.1"
torchmetrics = "^0.9.3"
hyperopt = "^0.2.7"

[tool.poetry.dev-dependencies]
black = {extras = ["jupyter"], version = "^22.8.0"}
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.950"
types-PyYAML = "^6.0.7"
pandas-profiling = "^3.2.0"
graphframes-stubs = "^0.8.0"
scalene = "1.5.8"

[tool.poetry.extras]
geometric = [
  "torch-scatter", 
  "torch-sparse", 
  "torch-cluster", 
  "torch-spline-conv", 
  "torch-geometric", 
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
